init UserProfileStoraget.class
